requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.1.0
forex-python==1.8
idna==3.10
inflection==0.5.1
ipython==8.28.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
stripe==11.1.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
